configurations refactoring: validation extracted from loader, data types extracted from base.py. create or update from json.
