+++ tests/test_configurations/test_loader.py
@@ -17,6 +17,7 @@
 #
 # ------------------------------------------------------------------------------
 
+
 """This module contains the tests for the aea.configurations.loader module."""
 import os
 from pathlib import Path
@@ -28,7 +29,8 @@ import yaml
 
 import aea
 from aea.configurations.base import PackageType, ProtocolSpecification
-from aea.configurations.loader import ConfigLoader, make_jsonschema_base_uri
+from aea.configurations.loader import ConfigLoader
+from aea.configurations.validation import make_jsonschema_base_uri
 from aea.protocols.generator.common import load_protocol_specification
 
 from tests.conftest import protocol_specification_files
